Update documentation on running running
The new documentation addersses aspects of the previous version which were
confusing.

A new option for running a model is provided in which the model can be
run from a script. In order to aid this functinoality, keyword arguments
were added to the run_module function. All command line flags can now
be directly entered as keyword arguments.

I also refocues the building a model section to highlight two core elements:
the config file and the dataset. I do this since these are two of the flags
which can be specified when running the model.