Separate user and developer docs
The previous documentation included instructions for installing/running
ZEN-garden as a user and developer in the same files. This led to lengthy
files and also required readers to skip around relevant sections.

I now created a new section "developer guide" which puts all the information
targeted at developers in one place. The previous "user guide" is now
exclusively focused on users which run ZEN-garden as a black box.