Add minimum full load hours constraint
This pull request adds a new conversion technology constraint. The
constraint requires conversion technologies to meet a minimum number of
full load hours per year. This is helpful for ensuring that technolgies
which can only be operated in a baseload fashion (e.g. nuclear) do not
exhibit ramping behavior.

The constraint primarily is documented in the docstring. This docstring
is then read into the documentation using the new Sphinx extension that
I programmed recently. I created a new test (test_1i) to test the
constraint and ensure that it keeps working in the future.

Thank you Jacob for the review and recommendations.