Modify post-processing for two-phase simulations
The post-processing script is modified to handle outputs from two-phase simulations. Results from the operation phase are appended to the existing h5 files created in the capacity planning phase. Only primal and dual variables of the operation phase can be saved. Variable names from the operations phase are modified with the suffix `_operation` to destinguish them from variables in the capacity planning phase.

For primal varaibles, a new solver option `selected_saved_variables_operation` is added. This controls which primal variables are saved from the operation phase. It works analagously to the previous option 'selected_saved_variables'. For duals, the same variables are saved in the operation phase as in the capacity phase. This is set by the solver option `selected_saved_duals`.