Improve Documentation of Default Configurations
I improved the documentation of the default configuration code.
This included writting a docstring for the module, and writting better
docstrings for some of the classes within the module. The docstrings
now also clearly show inheritance.

In addition to the docstrings, I restructured the Sphynx documentation
a bit to allow easier custom-configuration of the API docs. In
particular, custom automodule scripts can now be placed in the
`files\api\modules` folder and referenced as demonstrated in the
`files\api\models.rst`. This allows authors of the documentation to,
for example, specify individual flags and options for different
classes being documented. I added some code to the Sphynx `conf.py`
and the `make.bat` so that files in the `files\api\generated` folder
get deleted in every "make clean" and thus regenerated when the
documentation is built. In contrast, custom files in `files\api\modules`
are kept. This enables us to not push the generated files while still
giving us flexibility to manually overwrite the api files.