Restructure docs
I restructured the "zen_garden_in_detail" section of the documentation.
We have recieved feedback that the docs, which comprehensive, are a
bit difficult to navigate.

To help the structure, I moved some of the longer feature explaanations
(scaling, unit commitment, time series aggregation) to the tutorial
section. This makes them easier to find for users, since it makes the
information clearly visible in the table of contents. In the long-run, I
hope to develop these sections into proper tutorials with examples.

The "zen_garden_in_detail" section is now designed more as a reference
catalogue for the input structures, configurations, and equations of the
zen-garden model.