Add documentation guide
This commit adds a documentation guide. The guide
explains the basic structure of the documentation and rules for
contributing to the documentation. It also includes templates for
docstrings of classes and methods. Finally, the docstring extension is
described, which allows users to insert text from docstrings into the
documentation. It also describes how to override autodocs from Sphinx.

In addition to adding documentation, I also modified the docstring
extension to handle google-style docstrings. I recommend using google
style docstrings in future work since they are more concise and easier
to read.