Modify bumpversion to pull request instead of push
The bumpversion command previously directly pushed to the main branch.
I want to protect the branch so that direct pushes are no longer
possible. This created an error in the old bumpversion.

The new bumpversion automatically creates a new branch, pushes that
branch, submits a pull request, and then merges that pull request
into main. This allows us to keep the automated bumpversion with the
new branch protections.