Add github actions & workflow and mermaid to docs
I added some class diagrams to the API reference of the documentation.
To do this, I am using the Sphinx mermaid extension. Ideally, these
diagrams would be accompanied by a description of each class.
Furthermore, a description of the structure of OptimizationSetup is
still missing. #901

Furthermore, I obtained feedback on typos and issues in the
documentation from a student that is learning ZEN-garden.This commit
fixes these typos. Thank you Valentin for your valuable feedback.

Next, I documented the github actions changes in a new section of
the documentation related to github actions.

Finally, I cleaned the contribution guide to include more details on
the ZEN-garden workflow. Close #896