Fix pull requests in bumpversion
Bumpversion action was failing to make create and automerge a pull
request. Bumpversion correctly created version tags; however, the
bumpversion.cfg and the pyproject.toml files were not being updated.
This lead to incorrect versioning and errors in subsequent bumpversion
runs.

This commit modifies the bumpversion to fix the error and adjusts the
pyproject.toml and the bumpversion.cfg to contain the correct version
number. Thank you to Lucas Krauter(LucK1Y) for the help in solving this
issue.