Split actions into pull request and push
The previous bumpversion does not work since
its pull request needs to run tests before auto-
merging. The action 'release' does not wait for
these tests, hence it runs before bumpversion has
been merged. This causes the release to fail since
there is no new version number.

To address this, I have separated the github actions
into those which should run directly on the pull
request and those which should run on the push (i.e.
the pull request merge). Now, the tests only run on
the pull request. These must pass for the pull request
to be allowed to merge. Then, bumpversion and release
only run on push. This forces release to wait for
bumpversion.