Create tag after bumpversion_final
The previous version of the Github actions pipeline did not create a
tag for each new version. This caused the release to fail, since no
tag was identified.

I created a new GitHub action job  `create_tag` that manually creates
a tag. It reads the current version number from the bumpversion commit
message and then pushes that tag to the bumpversion commit.