Fix 'build' badge failure.
The build badge was marked as failed due to a past failed version
of the Github actions job release-test-pypi. I am therefore uncommenting
this job to fix the badge.

I also changed the base branch of the pull request action to the github
context gibhub.ref. This makes it easier to adapt the code to a new
branch when testing in the future. The only branch-specific information
is now coded in the header and the release jobs