Fix coverage badge in github actions
The coverage badge no longer seems to work. One reason why this may
be is because the tests only run on pull_request rather than on
a push to the main. I fixed this by making the tests now also run
on pushes to the main branch. The coverage badge action now only
runs in push events.

I also undid a previous change to the pull request action which
caused an error.