Increase flexibility of the config flag
Previously, the config flag required that a proper path to the config
file (e.g. "./config.json") be specified. I added the flexibility to
also simply state the name of the config file (e.g. "config.json") if
it is in the current working directory. The "dataset" input currently
allows for this flexibility. This commit therefore aligns the allowed
input types for the config and dataset flags.

Closes #1089