Fix config path input handling
ZEN-garden previously produced an error when the config file path
was not in the current working directory. This bug has now been
resolved.

Additionally, this commit fixes several additional issues related
to the file paths handling in ZEN-garden. First, ZEN-garden
changed the working directory to the directory of
the config file and did not restore the original working directory at
the end of the program. This hidden directory change can be confusing
to users. All directory changes in ZEN-garden have now been eliminated.
Next, the dataset path previously needed to be expressed relative to
the config path even though this was not documented. This commit makes
the dataset path relative to the current working directory, thus making
the two independent. Finally, the test cases has been reformulated
to match the changes to the input file paths. The new
test code gives full file paths for all datasets.

Closes #1099