Fix bug in folder_output configuration
When folder_output was specified, some output files were still saved
in the default location "./outputs". This had two reasons: (1) the
solver_dir was not adjusted to match folder_output, and (2) the
postprocessing changed incorrectly the folder_output settings when
converting from absolute to relative paths. The changes in this commit
address these bugs so that folder_output now properly sets the directory
in which all outputs are saved. The test cases now use this feature
to manually select the folder in which outputs are saved.