Fix bug in system.allow_investment
The setting system.allow_investment was being overwritten by
code for the setting system.include_operaton_only_phase. Investment
was therefore always being allowed in one-phase simulations, regardless
of what was specified in the system.allow_investment. This commit
fixes this bug.

Furthermore, the previous test for system.allow_investment was not
effective since the test result was the same regardless of whether
system.allow_investment was true or false. This commit introduces
a new test scenario that remedies this issue.